gynx_constants -> config に変更

+++ supabase/rls/test/t_posts_test.dart
@@ -1,5 +1,5 @@
+import 'package:config/config.dart';
 import 'package:faker/faker.dart';
-import 'package:gynx_constants/gynx_constants.dart';
 import 'package:supabase/supabase.dart';
 import 'package:test/test.dart';
 
@@ -35,10 +35,7 @@ void main() {
     });
 
     test('Select X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .select()
-          .eq('user_id', otherUserId);
+      final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
       expect(result, isEmpty);
     });
 
@@ -54,11 +51,7 @@ void main() {
     });
 
     test('Delete X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .delete()
-          .eq('user_id', otherUserId)
-          .select();
+      final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
       expect(result, isEmpty);
     });
   });
@@ -98,10 +91,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', anonymousUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', anonymousUserId);
         expect(result, isNotEmpty);
       });
 
@@ -117,11 +107,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', anonymousUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', anonymousUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -141,10 +127,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -161,11 +144,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -213,8 +192,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result =
-            await supabaseClient.from(tableName).select().eq('user_id', userId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', userId);
         expect(result, isNotEmpty);
       });
 
@@ -230,11 +208,7 @@ void main() {
       });
 
       test('Delete O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', userId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', userId).select();
         expect(result, isNotEmpty);
       });
     });
@@ -254,10 +228,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -273,11 +244,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isEmpty);
       });
     });
