+++ supabase/rls/test/t_user_follows_test.dart
@@ -1,5 +1,5 @@
+import 'package:config/config.dart';
 import 'package:faker/faker.dart';
-import 'package:gynx_constants/gynx_constants.dart';
 import 'package:supabase/supabase.dart';
 import 'package:test/test.dart';
 
@@ -39,10 +39,7 @@ void main() {
     });
 
     test('Select X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .select()
-          .eq('user_id', otherUserId1);
+      final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId1);
       expect(result, isEmpty);
     });
 
@@ -58,11 +55,7 @@ void main() {
     });
 
     test('Delete X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .delete()
-          .eq('user_id', otherUserId1)
-          .select();
+      final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId1).select();
       expect(result, isEmpty);
     });
   });
@@ -102,10 +95,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', anonymousUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', anonymousUserId);
         expect(result, isNotEmpty);
       });
 
@@ -121,11 +111,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', anonymousUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', anonymousUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -145,10 +131,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId1);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId1);
         expect(result, isNotEmpty);
       });
 
@@ -164,11 +147,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId1)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId1).select();
         expect(result, isEmpty);
       });
     });
@@ -208,8 +187,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result =
-            await supabaseClient.from(tableName).select().eq('user_id', userId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', userId);
         expect(result, isNotEmpty);
       });
 
@@ -225,11 +203,7 @@ void main() {
       });
 
       test('Delete O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', userId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', userId).select();
         expect(result, isNotEmpty);
       });
     });
@@ -249,10 +223,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId1);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId1);
         expect(result, isNotEmpty);
       });
 
@@ -268,11 +239,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId1)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId1).select();
         expect(result, isEmpty);
       });
     });
