+++ supabase/rls/test/t_user_post_favorites_test.dart
@@ -1,5 +1,5 @@
+import 'package:config/config.dart';
 import 'package:faker/faker.dart';
-import 'package:gynx_constants/gynx_constants.dart';
 import 'package:supabase/supabase.dart';
 import 'package:test/test.dart';
 
@@ -45,10 +45,7 @@ void main() {
     });
 
     test('Select X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .select()
-          .eq('user_id', otherUserId);
+      final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
       expect(result, isEmpty);
     });
 
@@ -64,11 +61,7 @@ void main() {
     });
 
     test('Delete X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .delete()
-          .eq('user_id', otherUserId)
-          .select();
+      final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
       expect(result, isEmpty);
     });
   });
@@ -108,10 +101,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', anonymousUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', anonymousUserId);
         expect(result, isNotEmpty);
       });
 
@@ -127,11 +117,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', anonymousUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', anonymousUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -151,10 +137,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -170,11 +153,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -214,8 +193,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result =
-            await supabaseClient.from(tableName).select().eq('user_id', userId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', userId);
         expect(result, isNotNull);
       });
 
@@ -231,11 +209,7 @@ void main() {
       });
 
       test('Delete O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', userId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', userId).select();
         expect(result, isNotEmpty);
       });
     });
@@ -255,10 +229,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -274,11 +245,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isEmpty);
       });
     });
