+++ supabase/rls/test/t_user_timelines_test.dart
@@ -1,5 +1,5 @@
+import 'package:config/config.dart';
 import 'package:faker/faker.dart';
-import 'package:gynx_constants/gynx_constants.dart';
 import 'package:supabase/supabase.dart';
 import 'package:test/test.dart';
 
@@ -46,10 +46,7 @@ void main() {
     });
 
     test('Select X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .select()
-          .eq('user_id', otherUserId);
+      final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
       expect(result, isEmpty);
     });
 
@@ -65,11 +62,7 @@ void main() {
     });
 
     test('Delete X', () async {
-      final result = await supabaseClient
-          .from(tableName)
-          .delete()
-          .eq('user_id', otherUserId)
-          .select();
+      final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
       expect(result, isEmpty);
     });
   });
@@ -122,10 +115,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', anonymousUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', anonymousUserId);
         expect(result, isNotEmpty);
       });
 
@@ -141,11 +131,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', anonymousUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', anonymousUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -166,10 +152,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -185,11 +168,7 @@ void main() {
       });
 
       test('Delete X', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isEmpty);
       });
     });
@@ -240,8 +219,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result =
-            await supabaseClient.from(tableName).select().eq('user_id', userId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', userId);
         expect(result, isNotEmpty);
       });
 
@@ -257,11 +235,7 @@ void main() {
       });
 
       test('Delete O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', userId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', userId).select();
         expect(result, isNotEmpty);
       });
     });
@@ -299,10 +273,7 @@ void main() {
       });
 
       test('Select O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .select()
-            .eq('user_id', otherUserId);
+        final result = await supabaseClient.from(tableName).select().eq('user_id', otherUserId);
         expect(result, isNotEmpty);
       });
 
@@ -318,11 +289,7 @@ void main() {
       });
 
       test('Delete O', () async {
-        final result = await supabaseClient
-            .from(tableName)
-            .delete()
-            .eq('user_id', otherUserId)
-            .select();
+        final result = await supabaseClient.from(tableName).delete().eq('user_id', otherUserId).select();
         expect(result, isNotEmpty);
       });
     });
